refactor codebase for AssignFamily
