+++ client/src/components/NewMateCard/NewMateCard.css
