+++ client/src/index.tsx
@@ -5,7 +5,7 @@ import React from "react";
 import ReactDOM from "react-dom";
 import { BrowserRouter as Router } from "react-router-dom";
 
-import App from "./components/App/App";
+import App from "./pages/App/App";
 import reportWebVitals from "./reportWebVitals";
 
 const client = new ApolloClient({
